requirements.txt
asgiref==3.8.0
Babel==2.14.0
Django==4.2.11
django-money==3.4.1
djangorestframework==3.15.1
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
py-moneyed==3.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1